requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
dateparser==1.3.0
frozenlist==1.8.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
propcache==0.4.1
pycryptodome==3.23.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python_binance==1.0.35
pytz==2026.1.post1
regex==2026.2.28
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
six==1.17.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
websockets==16.0
yarl==1.23.0
